refactor: more Firefox test changes
