+++ src/api/component.js
@@ -8,9 +8,9 @@ import {
 import { customElementsV0, reflect } from '../util/support';
 import data from '../util/data';
 import debounce from '../util/debounce';
-import empty from '../util/empty';
 import getAllKeys from '../util/get-all-keys';
 import getOwnPropertyDescriptors from '../util/get-own-property-descriptors';
+import syncPropToAttr from '../util/sync-prop-to-attr';
 
 const setElementAsDefined = elem => elem.setAttribute('defined', '');
 
@@ -147,41 +147,11 @@ Component.updated = function updated(elem, prev) {
   return false;
 };
 
-function getPropData(elem, name) {
-  const namespace = `api/property/${typeof name === 'symbol' ? String(name) : name}`;
-  return data(elem, namespace);
-}
-
-// taken from props-init.js TODO split into helper
-function getInitialValue(elem, name, opts) {
-  return typeof opts.initial === 'function' ? opts.initial(elem, { name }) : opts.initial;
-}
-
 function syncPropsToAttrs(elem) {
   const props = elem.constructor.props;
   Object.keys(props).forEach((propName) => {
-    const attributeName = data(elem, 'propertyLinks')[propName];
-    if (!attributeName) return false;
-
-    // let shouldRemoveAttribute = false;
-    const propData = getPropData(elem, propName);
     const prop = props[propName];
-    // console.log(prop, prop.internalValue);
-    let syncAttrValue = propData.lastAssignedValue;
-    // console.log('a:', syncAttrValue);
-    if (empty(syncAttrValue) && prop.initial) {
-      syncAttrValue = getInitialValue(elem, propName, prop);
-    }
-    if (!empty(syncAttrValue) && prop.serialize) {
-      syncAttrValue = prop.serialize(syncAttrValue);
-    }
-    // console.log('b:', syncAttrValue);
-    if (!empty(syncAttrValue)) {
-      // console.log(`Setting ${propName} attr to ${syncAttrValue}`);
-      propData.syncingAttribute = true;
-      // propData.settingAttribute = false;
-      elem.setAttribute(attributeName, syncAttrValue);
-    }
+    syncPropToAttr(elem, prop, propName, true);
   });
 }
 
@@ -190,7 +160,6 @@ Component.prototype = Object.create(HTMLElement.prototype, {
   connectedCallback: {
     configurable: true,
     value() {
-      // console.log('connectedCallback', window.navigator.userAgent);
       syncPropsToAttrs(this);
       callConnected(this);
       setElementAsDefined(this);
@@ -209,10 +178,8 @@ Component.prototype = Object.create(HTMLElement.prototype, {
   attributeChangedCallback: {
     configurable: true,
     value(name, oldValue, newValue) {
-      // console.log('attributeChangedCallback', name, oldValue, newValue);
       const { attributeChanged, observedAttributes } = this.constructor;
       const propertyName = data(this, 'attributeLinks')[name];
-      // console.log('propertyName', propertyName);
 
       // In V0 we have to ensure the attribute is being observed.
       if (customElementsV0 && observedAttributes.indexOf(name) === -1) {
@@ -227,14 +194,11 @@ Component.prototype = Object.create(HTMLElement.prototype, {
         // chance to run when you set an attribute, it then sets a property and
         // then that causes the attribute to be set again.
         if (propData.syncingAttribute) {
-          // console.log('syncingAttribute');
           propData.syncingAttribute = false;
         } else {
           // Sync up the property.
           const propOpts = this.constructor.props[propertyName];
-          // console.log('propOpts', propOpts);
           propData.settingAttribute = true;
-          // console.log('setting prop to ', (newValue !== null && propOpts.deserialize ? propOpts.deserialize(newValue) : newValue));
           this[propertyName] = newValue !== null && propOpts.deserialize ? propOpts.deserialize(newValue) : newValue;
         }
       }
@@ -249,7 +213,6 @@ Component.prototype = Object.create(HTMLElement.prototype, {
   createdCallback: {
     configurable: true,
     value() {
-      // console.log('createdCallback');
       callConstructor(this);
     },
   },
@@ -258,7 +221,6 @@ Component.prototype = Object.create(HTMLElement.prototype, {
   attachedCallback: {
     configurable: true,
     value() {
-      // console.log('attachedCallback', window.navigator.userAgent);
 
       syncPropsToAttrs(this);
