+++ src/util/get-initial-value.js
@@ -1,0 +1,3 @@
+export default function getInitialValue(elem, name, opts) {
+  return typeof opts.initial === 'function' ? opts.initial(elem, { name }) : opts.initial;
+}
